change c++ declrs to floats for now

+++ src/dios-egraphs/Diospyros/llvm-tests/a.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
-int a_in[] = {1, 2, 3, 4};
-int b_in[] = {5, 6, 7, 8};
+float a_in[] = {1, 2, 3, 4};
+float b_in[] = {5, 6, 7, 8};
 
 int main(int argc, char **argv) {
 //  return argc + 5;
-  int c_out[4];
+  float c_out[4];
   c_out[0] = a_in[0] + b_in[0];
   c_out[1] = a_in[1] + b_in[1];
   c_out[2] = a_in[2] + b_in[2];
