+++ src/dios-egraphs/Diospyros/llvm-tests/add_mult.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
-int a_in[] = {1, 2, 3, 4};
-int b_in[] = {2, 3, 4, 5};
+float a_in[] = {1, 2, 3, 4};
+float b_in[] = {2, 3, 4, 5};
 
 int main(int argc, char **argv) {
-    int d_out[4];
+    float d_out[4];
     d_out[0] = a_in[0] + b_in[0];
     d_out[1] = a_in[1] * b_in[1];
     d_out[2] = a_in[2] + b_in[2];
