+++ src/dios-egraphs/Diospyros/llvm-tests/mac.c
@@ -5,7 +5,7 @@ float b_in[] = {2, 3, 4, 5};
 float c_in[] = {3, 4, 5, 6};
 
 int main(int argc, char **argv) {
-    int d_out[4];
+    float d_out[4];
     d_out[0] = a_in[0] + (b_in[0] * c_in[0]);
     d_out[1] = a_in[1] + (b_in[1] * c_in[1]);
     d_out[2] = a_in[2] + (b_in[2] * c_in[2]);
